Added control panel for game actions. Added UI enhancements to buttons & moved elements around. Added RNG to water collection scene. Added money status to water collection scene. Adjusted range on the RNG for the water shop page. Added New Offer button and logic to shop menu.

+++ Unity/Live for Money/Assets/Scripts/PlayerScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerScript : MonoBehaviour
 {
-    public static int waterQuantity;
+    public static double waterQuantity;
     public static double moneyQuantity;
     // Start is called before the first frame update
     void Start()
